requirements.txt
Babel==2.6.0
backcall==0.1.0
decorator==4.4.0
Django==2.2
django-ckeditor==5.6.1
django-extensions==2.1.6
django-filter==2.1.0
django-js-asset==1.2.2
django-phonenumber-field==2.3.1
django-registration-redux==2.6
djangorestframework==3.9.2
easy-thumbnails==2.6
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
jsonfield==2.0.2
Markdown==3.1
parso==0.4.0
pexpect==4.7.0
phonenumbers==8.10.10
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
